In this update, we have converted many redundant Python configurations into more stable JSON configurations, while also achieving a preliminary implementation of Chaquopy integration.

+++ scripts/build_modules.py
@@ -2,17 +2,18 @@ import os
 import subprocess
 import shutil
 
-def build_modules(dir_path, exe_dir, python_exe):
+def build_modules(dir_path, exe_dir, python_exe, is_compile=True):
     output_path = os.path.join(exe_dir, os.path.basename(dir_path))
     if not os.path.exists(output_path):
         os.makedirs(output_path)
     for module in os.listdir(dir_path):
         path = os.path.join(dir_path, module)
         if os.path.isdir(path):
-            build_modules(path, output_path, python_exe)
+            build_modules(path, output_path, python_exe, is_compile)
         elif module.endswith(".py"):
             subprocess.run([python_exe, "-m", "python_minifier", path, "--output", os.path.join(output_path, module)], cwd=exe_dir)
-            subprocess.run([python_exe, "-m", "compileall", "-b", os.path.join(output_path, module)], cwd=exe_dir)
-            os.remove(os.path.join(output_path, module))
+            if is_compile:
+                subprocess.run([python_exe, "-m", "compileall", "-b", os.path.join(output_path, module)], cwd=exe_dir)
+                os.remove(os.path.join(output_path, module))
         else:
             shutil.copy2(path, output_path)
